# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers/me/profile": {
      "get": {
        "operationId": "CustomersController_profile",
        "summary": "[CUSTOMER APP] Profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "patch": {
        "operationId": "CustomersController_updateProfile",
        "summary": "[CUSTOMER APP] Update Profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/customers/me/preferences": {
      "get": {
        "operationId": "CustomersController_preferences",
        "summary": "[CUSTOMER APP] Preferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "patch": {
        "operationId": "CustomersController_updatePreferences",
        "summary": "[CUSTOMER APP] Update Preferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/customers/me/security/sessions": {
      "get": {
        "operationId": "CustomersController_securitySessions",
        "summary": "[CUSTOMER APP] List refresh sessions owned by the customer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerSecuritySessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/customers/me/security/sessions/{id}": {
      "delete": {
        "operationId": "CustomersController_revokeSecuritySession",
        "summary": "[CUSTOMER APP] Revoke an owned refresh session",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Refresh session revoked"
          },
          "404": {
            "description": "Refresh session not found"
          }
        },
        "tags": [
          "Customer App - Profile"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
...
    "/api/v1/cart": {
      "get": {
        "operationId": "CartsController_detail",
        "summary": "[CUSTOMER APP] Detail",
        "parameters": [
          {
            "name": "restaurantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Restaurant whose active cart should be returned; defaults to the active restaurant"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Carts"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "delete": {
        "operationId": "CartsController_clear",
        "summary": "[CUSTOMER APP] Clear the authenticated customer cart",
        "parameters": [
          {
            "name": "restaurantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customer App - Carts"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "operationId": "CartsController_add",
        "summary": "[CUSTOMER APP] Add",
        "parameters": [
          {
            "name": "restaurantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Restaurant that owns the menu item and cart; defaults to the active restaurant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Carts"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/cart/items/{id}": {
      "patch": {
        "operationId": "CartsController_update",
        "summary": "[CUSTOMER APP] Update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Carts"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "delete": {
        "operationId": "CartsController_remove",
        "summary": "[CUSTOMER APP] Remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Carts"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
...
    "/api/v1/checkout": {
      "post": {
        "operationId": "CheckoutController_checkout",
        "summary": "[CUSTOMER APP] Checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Idempotency key request conflict"
          }
        },
        "tags": [
          "Customer App - Checkout"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
...
  },
  "info": {
    "title": "La Favola Pizza Restaurant API",
    "description": "Routes are grouped by audience. Customer app developers should use sections prefixed \"Customer App\" and \"Audience: Public / Customer App\". Administrative and staff operations declare their required roles in their Audience tag and x-required-roles metadata.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateCustomerProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCustomerPreferencesDto": {
        "type": "object",
        "properties": {}
      },
      "CustomerSecuritySessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "revoked": {
            "type": "boolean"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Present only after the refresh session has been revoked"
          }
        },
        "required": [
          "id",
          "createdAt",
          "expiresAt",
          "revoked"
        ]
      },
...
      "AddCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {}
      },
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "idempotencyKey"
        ]
      },
      "AppliedPromotionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "discountMinor": {
            "type": "number"
          },
          "deliveryDiscountMinor": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "discountMinor",
          "deliveryDiscountMinor"
        ]
      },
      "CheckoutResultDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "amountMinor": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "subtotalMinor": {
            "type": "number"
          },
          "optionChargesMinor": {
            "type": "number"
          },
          "deliveryFeeMinor": {
            "type": "number"
          },
          "taxMinor": {
            "type": "number"
          },
          "promotionDiscountMinor": {
            "type": "number"
          },
          "couponDiscountMinor": {
            "type": "number"
          },
          "loyaltyDiscountMinor": {
            "type": "number"
          },
          "deliveryDiscountMinor": {
            "type": "number"
          },
          "appliedPromotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppliedPromotionDto"
            }
          },
          "estimatedDeliveryAt": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedReadyAt": {
            "format": "date-time",
            "type": "string"
          },
          "serverNow": {
            "format": "date-time",
            "type": "string"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "delivery",
              "pickup"
            ]
          }
        },
        "required": [
          "orderId",
          "orderNumber",
          "status",
          "paymentStatus",
          "amountMinor",
          "currency",
          "subtotalMinor",
          "optionChargesMinor",
          "deliveryFeeMinor",
          "taxMinor",
          "promotionDiscountMinor",
          "couponDiscountMinor",
          "loyaltyDiscountMinor",
          "deliveryDiscountMinor",
          "appliedPromotions",
          "serverNow",
          "orderType"
        ]
      },
...
    }
  },
  "security": [
    {
      "JWT-auth": []
    }
  ]
}